requirements.txt
django
gunicorn
django-heroku
django-webpack-loader
psycopg2
djangorestframework